updated docs with new result obj
